Day16!
This is definitely one of the problems that presents you with a part1
that is intentionally meant to mislead us into a naive implementation.

But when the message is large and the offset we seek is more than halfway
through the input message (repeated 10000 times,) we are able to exploit
the multiplier pattern in the following way:

At input index N, the pattern will be (N - 1) x 0, N x 1, N x 0, N x -1

If N > len(input) / 2, then the pattern will be only 0's and 1's across
the entire input. We can then just sum the 1's ONCE to get the first
digit at N, then just iterate once more from N to the end, subtracting
each input value at N from the sum to get what would have been the sum
of that next digit if we added from input[N+1:] all over again.

Put another way, in Pythonic pseudo-code, let N be the offset, each
step of the phase computation is:

if (N > len(input) / 2)
  output[N] = sum(input[N:])
  output[N+1] = output[N] - input[N+1]

So each phase computation is now two addition loops over a maximum of
3.25M elements, and the given offset tends towards the last 3/4 of the
input size set, somewhere between 5.5M and 6.5M, so it is more
realistcally a maximum of 1M iterations x 2 x 100 = 200M iterations of
very simple operations to figure out part two.

It could probably be optimized depending on vector growth performance in
C++, but it runs pretty quickly!

How the hell would this be visualized tho?