Day17 solution!
What a difficult puzzle. There were many algorithms needing
implementation here: not only did the intcode processor need to work
exactly as specified (i had a subtle issue in incrementing the pc that
was fixed an earlier commit,) but also needed to process the output,
understanding exactly how the map is formatted, then deconstructing that
map output into a set of instructions, and then compressing those
instructions.

My naive first approach was a BFS from the origin to the end, which at a
peak had about 8,000 possibilities and was very slow to munge. Only then
did I realize that the optimum path will explicitly be the one with the
fewest turns, or, the longest runs. Once that realization was made, that
algorith was easier. Then implementing the compression algorith was
tricky, as well, and required a little bit of cheating on my part to see
what the right approach was: knowing that it had to start on a turn and
end on a distance made it much easier to figure out.

Phew. This was a tough one!